Day 05 Solving without Sorting
My child is trying to solve AoC without storing data. This means saving the seats and sorting, the obvious solution is out. I am not sure this solution works for him, but it's an interesting twist. This version shows an implementation that does not require sorting to find the seat. This is probably not entirely true. The seats are sorted as they are encountered, but it does eliminate the need to save all of the seats.