Day 05 Rewrite
This solution keeps track of the first and last seats as well as the sum of all seat numbers encountered. We can compute the expected sum of all seats and subtract the observed sum to find the empty seat.

I've also simplified the conversion of the boarding pass to the seat id.

Source: Facebook AP CSA Austin McGowan