Day 16: Ticket Translation
Fun but complex. There was no foolish wand-waving or silly incantations just processing data. I left some debug statements in place.

There is code to reduce the field ranges into a minimal set of valid field values that I thought might be useful but, a) it builds the set backwards and b) there is a warning for a condition I don't understand. So though it might be useful to more quickly eliminate bad tickets it's not being used.

I have a separate step to reduce the fields down at the end. I think a better solution would be to keep the fields as maps and reduce it as you go. This would make the code more readable and reduce processing during and at the end of the run.