Day 24: Lobby Layout
I had to learn hex coordinate systems to solve this. I initially implemented it using a cube system but after reviewing reddit I switched to an offset system.

I had a lot of issues with this one and added a visualization to help me debug. The visualization is currently useless and I spent more time getting it working than it was worth (obviously). I ended up realizing the problem was (again) shallow copies of maps.